Handle erlang.mk dep_depname_commit format
In erlang.mk one can specify a dep with repo type, url and commit.
But if it is a supported dep, it is enough to mention the commit
hash or name of the branch or the tag. Like

dep_myrepo = git git@github.com:jonasrichard/myrepo.git master
dep_jsx_commit = 2.8.0

In that case the dep variable will be suffixed with _commit.
Now elvis fails on them, since it is not a proper repo. This
commit fixes that, by filtering out the deps having _commit
suffix.