Rock in parallel
Implement parallel processing of the files. Option 'parallel'
takes pos integer as a number of processes allowed to run in
parallel, default to 1, to keep current behaviour.

Drop most of the file section from the result, and split loading
phase to reduce memory consumption and speed up gathering of the
results.